Some fixes to low-level network routines.
Also, in general, change things so that the server does not asynchronously
send stuff to the client; everything from server to client is in
response to a message from the client.

The clients poll for state, and also updated state is sent back when
the client sends a message like "zone heating enable".